Create RFID-Reader-Logger
Created a new file containing all the required RFID reading and logging code. Still needs servo actuation code to be triggered in main as well as any buzzer or led control. 

Logging of unauthorized users has yet to be tested.

Use on battery or outlet power has yet to be tested